#62 Implementacion-politicas

+++ app/Http/Controllers/SubjectController.php
@@ -14,14 +14,17 @@ class SubjectController extends Controller
     ];
     public function index()
     {
+        $this->authorize('viewAny', Subject::class);
         return Subject::all();
     }
     public function show(Subject $subject)
     {
+        $this->authorize('view', $subject);
         return response()->json(new SubjectResource($subject),200);
     }
     public function store(Request $request)
     {
+        $this->authorize('create', Subject::class);
         $request->validate([
             'name' => 'required|string',
             'level' => 'required|string',
@@ -32,6 +35,7 @@ class SubjectController extends Controller
     }
     public function update(Request $request, Subject $subject)
     {
+        $this->authorize('update',$subject);
         $request->validate([
             'level' => 'required|string',
         ],self::$messages);
@@ -40,6 +44,7 @@ class SubjectController extends Controller
     }
     public function delete(Request $request, Subject $subject)
     {
+        $this->authorize('delete',$subject);
         $subject->delete();
         return response()->json(null, 204);
     }
